// integrations/supabase/types.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "13.0.5"
  }
  public: {
    Tables: {
// ...
      Company: {
        Row: {
          address_id: number | null
          aggregated_late_payment_days: number | null
          amount_overdue: number | null
          created_at: string | null
          credit_limit: number | null
          current_exposure: number | null
          detected_last_checked: string | null
          detected_profile_id: string | null
          detected_review_status: string | null
          detected_risk_category: string | null
          detected_risk_label: string | null
          id: number
          kyb_effective_date: string | null
          kyb_status: Database["public"]["Enums"]["kyb_status"] | null
          name: string
          payment_terms: string | null
          risk_rating: string | null
          total_MT_loaded: number | null
          total_MT_signed: number | null
          total_traded_volume: number | null
          trade_credit_limit: number | null
          trader_relationship_owner:
            | Database["public"]["Enums"]["trader_relationship_owner"]
            | null
        }
        Insert: {
          address_id?: number | null
          aggregated_late_payment_days?: number | null
          amount_overdue?: number | null
          created_at?: string | null
          credit_limit?: number | null
          current_exposure?: number | null
          detected_last_checked?: string | null
          detected_profile_id?: string | null
          detected_review_status?: string | null
          detected_risk_category?: string | null
          detected_risk_label?: string | null
          id?: number
          kyb_effective_date?: string | null
          kyb_status?: Database["public"]["Enums"]["kyb_status"] | null
          name: string
          payment_terms?: string | null
          risk_rating?: string | null
          total_MT_loaded?: number | null
          total_MT_signed?: number | null
          total_traded_volume?: number | null
          trade_credit_limit?: number | null
          trader_relationship_owner?:
            | Database["public"]["Enums"]["trader_relationship_owner"]
            | null
        }
        Update: {
          address_id?: number | null
          aggregated_late_payment_days?: number | null
          amount_overdue?: number | null
          created_at?: string | null
          credit_limit?: number | null
          current_exposure?: number | null
          detected_last_checked?: string | null
          detected_profile_id?: string | null
          detected_review_status?: string | null
          detected_risk_category?: string | null
          detected_risk_label?: string | null
          id?: number
          kyb_effective_date?: string | null
          kyb_status?: Database["public"]["Enums"]["kyb_status"] | null
          name?: string
          payment_terms?: string | null
          risk_rating?: string | null
          total_MT_loaded?: number | null
          total_MT_signed?: number | null
          total_traded_volume?: number | null
          trade_credit_limit?: number | null
          trader_relationship_owner?:
            | Database["public"]["Enums"]["trader_relationship_owner"]
            | null
        }
        Relationships: [
          {
            foreignKeyName: "Company_address_id_fkey"
            columns: ["address_id"]
            isOneToOne: false
            referencedRelation: "Company_address"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
      order: {
        Row: {
          allocated_quantity_mt: number | null
          buy_downpayment_amount: number | null
          buy_downpayment_due_date: string | null
          buy_downpayment_invoice: Json | null
          buy_downpayment_paid_date: string | null
          buy_price: number | null
          buyer: string | null
          commodity_type:
            | Database["public"]["Enums"]["commodity_type_enum"]
            | null
          created_at: string | null
          delete_reason: string | null
          deleted_at: string | null
          id: string
          isri_grade: Database["public"]["Enums"]["isri_grade_enum"] | null
          loading_date: string | null
          margin: number | null
          metal_form: Database["public"]["Enums"]["metal_form_enum"] | null
          partial_shipment_allowed: boolean | null
          product_details: string | null
          purchase_order_url: string | null
          sales_order_sign_date: string | null
          sales_order_url: string | null
          sell_downmapayment_invoice: Json | null
          sell_downpayment_amount: number | null
          sell_downpayment_due_date: string | null
          sell_downpayment_paid_date: string | null
          sell_price: number | null
          seller: string | null
          ship_from: string | null
          ship_to: string | null
          signed_purchase_order_url: string | null
          signed_sales_order_url: string | null
          status: string | null
          transaction_type: string | null
        }
        Insert: {
          allocated_quantity_mt?: number | null
          buy_downpayment_amount?: number | null
          buy_downpayment_due_date?: string | null
          buy_downpayment_invoice?: Json | null
          buy_downpayment_paid_date?: string | null
          buy_price?: number | null
          buyer?: string | null
          commodity_type?:
            | Database["public"]["Enums"]["commodity_type_enum"]
            | null
          created_at?: string | null
          delete_reason?: string | null
          deleted_at?: string | null
          id: string
          isri_grade?: Database["public"]["Enums"]["isri_grade_enum"] | null
          loading_date?: string | null
          margin?: number | null
          metal_form?: Database["public"]["Enums"]["metal_form_enum"] | null
          partial_shipment_allowed?: boolean | null
          product_details?: string | null
          purchase_order_url?: string | null
          sales_order_sign_date?: string | null
          sales_order_url?: string | null
          sell_downmapayment_invoice?: Json | null
          sell_downpayment_amount?: number | null
          sell_downpayment_due_date?: string | null
          sell_downpayment_paid_date?: string | null
          sell_price?: number | null
          seller?: string | null
          ship_from?: string | null
          ship_to?: string | null
          signed_purchase_order_url?: string | null
          signed_sales_order_url?: string | null
          status?: string | null
          transaction_type?: string | null
        }
        Update: {
          allocated_quantity_mt?: number | null
          buy_downpayment_amount?: number | null
          buy_downpayment_due_date?: string | null
          buy_downpayment_invoice?: Json | null
          buy_downpayment_paid_date?: string | null
          buy_price?: number | null
          buyer?: string | null
          commodity_type?:
            | Database["public"]["Enums"]["commodity_type_enum"]
            | null
          created_at?: string | null
          delete_reason?: string | null
          deleted_at?: string | null
          id?: string
          isri_grade?: Database["public"]["Enums"]["isri_grade_enum"] | null
          loading_date?: string | null
          margin?: number | null
          metal_form?: Database["public"]["Enums"]["metal_form_enum"] | null
          partial_shipment_allowed?: boolean | null
          product_details?: string | null
          purchase_order_url?: string | null
          sales_order_sign_date?: string | null
          sales_order_url?: string | null
          sell_downmapayment_invoice?: Json | null
          sell_downpayment_amount?: number | null
          sell_downpayment_due_date?: string | null
          sell_downpayment_paid_date?: string | null
          sell_price?: number | null
          seller?: string | null
          ship_from?: string | null
          ship_to?: string | null
          signed_purchase_order_url?: string | null
          signed_sales_order_url?: string | null
          status?: string | null
          transaction_type?: string | null
        }
        Relationships: []
      }
// ...
      pricing_fixing: {
        Row: {
          bl_order_id: number | null
          created_at: string
          delete_reason: string | null
          deleted_at: string | null
          final_price: number | null
          final_price_currency: string | null
          fixed_at: string | null
          id: string
          level: string | null
          metal: Database["public"]["Enums"]["commodity_type_enum"]
          notes: string | null
          order_id: string | null
          premium_discount: number | null
          premium_discount_currency: string | null
          qp_end: string | null
          qp_start: string | null
          quantity_mt: number | null
          reference: Database["public"]["Enums"]["reference_type"]
          ticket_id: number | null
          updated_at: string
        }
        Insert: {
          bl_order_id?: number | null
          created_at?: string
          delete_reason?: string | null
          deleted_at?: string | null
          final_price?: number | null
          final_price_currency?: string | null
          fixed_at?: string | null
          id?: string
          level?: string | null
          metal: Database["public"]["Enums"]["commodity_type_enum"]
          notes?: string | null
          order_id?: string | null
          premium_discount?: number | null
          premium_discount_currency?: string | null
          qp_end?: string | null
          qp_start?: string | null
          quantity_mt?: number | null
          reference?: Database["public"]["Enums"]["reference_type"]
          ticket_id?: number | null
          updated_at?: string
        }
        Update: {
          bl_order_id?: number | null
          created_at?: string
          delete_reason?: string | null
          deleted_at?: string | null
          final_price?: number | null
          final_price_currency?: string | null
          fixed_at?: string | null
          id?: string
          level?: string | null
          metal?: Database["public"]["Enums"]["commodity_type_enum"]
          notes?: string | null
          order_id?: string | null
          premium_discount?: number | null
          premium_discount_currency?: string | null
          qp_end?: string | null
          qp_start?: string | null
          quantity_mt?: number | null
          reference?: Database["public"]["Enums"]["reference_type"]
          ticket_id?: number | null
          updated_at?: string
        }
        Relationships: [
          {
            foreignKeyName: "pricing_fixing_bl_order_id_fkey"
            columns: ["bl_order_id"]
            isOneToOne: false
            referencedRelation: "bl_order"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "pricing_fixing_order_id_fkey"
            columns: ["order_id"]
            isOneToOne: false
            referencedRelation: "order"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "pricing_fixing_ticket_id_fkey"
            columns: ["ticket_id"]
            isOneToOne: false
            referencedRelation: "ticket"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
    }
// ...
    Enums: {
// ...
      commodity_type_enum:
        | "Aluminium"
        | "Mixed metals"
        | "Zinc"
        | "Magnesium"
        | "Lead"
        | "Nickel/stainless/hi-temp"
        | "Copper"
        | "Brass"
        | "Steel"
        | "Iron"
// ...
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}